Update Biginteger with ulong.

+++ src/SoterDevice/Models/IEthereumTransaction.cs
@@ -23,8 +23,8 @@ namespace SoterDevice.Models
     public interface IEthereumTransaction : ITransaction
     {
         uint ChainId { get; }
-        BigInteger GasPrice { get; }
-        BigInteger GasLimit { get; }
+        ulong GasPrice { get; }
+        ulong GasLimit { get; }
         BigInteger Nonce { get; }
         byte[] Data { get; }
     }
